[loop-analysis] Shared input shape inference builds. Untested.
